fix(apisimp): guard page*limit arithmetic overflow in ReferenceAnnotationRest.list()
CodeQL flagged two high-severity arithmetic-overflow paths:
- page * limit: page has no @Max, so large values overflow int
- from + limit: same risk if from is near Integer.MAX_VALUE

Fix: cast to long before multiplication/addition; cast result back to int
after Math.min clamps it to [0, total]. Both casts are now safe because
the clamped value fits in an int (total is an int by construction).

Refs: APISIMP-REF-ANNOTATION-LIST-NO-PAGINATION / PR #2237 CodeQL gate

+++ backend/src/main/java/de/dlr/shepard/v2/references/resources/ReferenceAnnotationRest.java
@@ -160,8 +160,8 @@ public class ReferenceAnnotationRest {
     return gateAndDispatch(appId, AccessType.Read, sc, r -> {
       List<Map<String, Object>> rows = r.handler().listAnnotations(appId);
       int total = rows.size();
-      int from = Math.min(page * limit, total);
-      List<Map<String, Object>> slice = rows.subList(from, Math.min(from + limit, total));
+      int from = (int) Math.min((long) page * limit, (long) total);
+      List<Map<String, Object>> slice = rows.subList(from, (int) Math.min((long) from + limit, (long) total));
       return Response.ok(new PagedResponseIO<>(slice, total, page, limit)).build();
     });
   }
